LayoutTests/media/media-source/media-source-airplay-expected.txt


** create a MediaSource object, set as video.src
RUN(source = new MediaSource())
RUN(source.addEventListener('sourceopen', startLoad))
RUN(video.src = URL.createObjectURL(source))
EVENT( webkitplaybacktargetavailabilitychanged ) : event.availability = 'not-available'

** create a source buffer, add an init segment with 1 video track
RUN(sourceBuffer = source.addSourceBuffer('video/mock; codecs="mock"'))
RUN(init = makeAInit(100, [makeATrack(1, 'mock', TRACK_KIND.VIDEO)]))
RUN(sourceBuffer.appendBuffer(init))

EVENT(canplaythrough)

** simulate choosing a device from the menu
EVENT( webkitplaybacktargetavailabilitychanged ) : event.availability = 'available'

** change video.src to one that supports AirPlay
EXPECTED (video.webkitCurrentPlaybackTargetIsWireless == 'true') OK

END OF TEST
